fixes the multi autoload issue that is present in php verstions 5.3.0 and 5.3.2 where fully qualified class names would be loaded by everu autoloader even if they were already found (this bug is fixed in php 5.3.6). It fixes it by not storing class names as fully qualified. I think this is more in line with how php does string class names anyway (the reflection objects never have the beginning \).
Fixes a bug where Container::setPreferred() was being ignored if the class was not in the reflection instance. Adds Container::normalizeParam() so only one parameter of a method could be normalized at a time and then refactors Framework::handleController() to iterate through the params and do its thing (eg, check for forms) without having to normalize all the params and then iterate through them again.