lots of little fixes and updates, StdAutoloader now makes sure paths exist before trying to iterate over their contents. Fieldable instances can now be populated on creation by setting variables in a config file. Cacheable instances will now automatically have their cache loaded on creation if they have a cache instance set. Handling exceptions is now done by the exception short name, it wasn't clear on how to create a handle method of an exception that had a namespace, now an exception like \Foo\Bar\CheException can be handled by handleCheException(), while this is not ideal, because there could be a Che\CheException() it works for now and I'm not sure how common name clashes will be in the real world.
page.php no longer sets class=container, this was for convenience when using the Blueprint css framework, and that is too implementation specific, so I removed it.

Form field annotation was updated to provide a default value, and a submit button was placed in the common Form class, based on how I'm developing on it, I want this more than not, so for the few times I don't want the submit button, I can just set  = false; in the child class.

the Cacheable interface no longer requires killCache() or cacheParams() methods.