we are back in business, after making some changes that broke the entire framework months ago, it finally works again with the new changes. It still has some rough spots, but I can start developing with it again. Main changes:
command and endpoint are now 2 separate controller classes, so you would create a class that extended command to have a controller that handle a task or command line program, and extend endpoint to have a controller that handles a web request.

There is no longer an error controller, uncaught exceptions will now cause a filter event to be raised, which will give a chance to change how stuff will be outputted and everything, I think this is conceptually better, and easier to understand than trying to create an error class whose methods map to exceptions.

start classes no longer exist, use the startup events to initialize things now, this makes sense because when I added the configuration handlers it seemed strange to have 2 ways to configure everything.

Plugins can now have plugins, which can also have plugins, in one big recursive family